dom/websocket/tests/test_worker_websocket_https.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test that creating insecure websockets from https workers is not possible</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" language="javascript">

  onmessage = function(event) {
    is(event.data, "not created", "WebSocket object must not be created");
    SimpleTest.finish();
  };
  SimpleTest.waitForExplicitFinish();

</script>
</pre>
<iframe src="https://example.com/tests/dom/websocket/tests/websocket_worker_https.html"></iframe>
</body>
</html>
